+++ packages/spacecat-shared-content-client/CHANGELOG.md
@@ -1,3 +1,9 @@
+## [@adobe/spacecat-shared-content-client-v1.8.24](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-content-client-v1.8.23...@adobe/spacecat-shared-content-client-v1.8.24) (2026-05-14)
+
+### Bug Fixes
+
+* **deps:** update external fixes ([#1533](https://github.com/adobe/spacecat-shared/issues/1533)) ([0a3e2ab](https://github.com/adobe/spacecat-shared/commit/0a3e2abbbc5f58b5320518f7d596d4cef6271fa0))
+
 ## [@adobe/spacecat-shared-content-client-v1.8.23](https://github.com/adobe/spacecat-shared/compare/@adobe/spacecat-shared-content-client-v1.8.22...@adobe/spacecat-shared-content-client-v1.8.23) (2026-04-09)
 
 ### Bug Fixes